+++ datamodel/odata-v4/odata-v4-generator/src/test/resources/oDataGeneratorIntegrationTest/actionsAndFunctionsTest/output/testcomparison/namespaces/actionsandfunctions/FunctionResult.java
@@ -130,15 +130,15 @@ public class FunctionResult
         // simple properties
         {
             if (cloudSdkValues.containsKey("RequestId")) {
-                final Object value = cloudSdkValues.remove("RequestId");
-                if ((value == null)||(!value.equals(getRequestId()))) {
-                    setRequestId(((UUID) value));
+                final Object cloudSdkValue = cloudSdkValues.remove("RequestId");
+                if ((cloudSdkValue == null)||(!cloudSdkValue.equals(getRequestId()))) {
+                    setRequestId(((UUID) cloudSdkValue));
                 }
             }
             if (cloudSdkValues.containsKey("Message")) {
-                final Object value = cloudSdkValues.remove("Message");
-                if ((value == null)||(!value.equals(getMessage()))) {
-                    setMessage(((java.lang.String) value));
+                final Object cloudSdkValue = cloudSdkValues.remove("Message");
+                if ((cloudSdkValue == null)||(!cloudSdkValue.equals(getMessage()))) {
+                    setMessage(((java.lang.String) cloudSdkValue));
                 }
             }
         }