+++ datamodel/odata-v4/odata-v4-generator/src/test/resources/oDataGeneratorIntegrationTest/testService/output/testcomparison/namespaces/test/TestEntityCircularLinkChild.java
@@ -125,9 +125,9 @@ public class TestEntityCircularLinkChild
         // simple properties
         {
             if (cloudSdkValues.containsKey("KeyProperty")) {
-                final Object value = cloudSdkValues.remove("KeyProperty");
-                if ((value == null)||(!value.equals(getKeyProperty()))) {
-                    setKeyProperty(((java.lang.String) value));
+                final Object cloudSdkValue = cloudSdkValues.remove("KeyProperty");
+                if ((cloudSdkValue == null)||(!cloudSdkValue.equals(getKeyProperty()))) {
+                    setKeyProperty(((java.lang.String) cloudSdkValue));
                 }
             }
         }